Moves ApproGraph to components to use it more widely
