+++ frontend/src/components/ApproGraph.vue
